backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
langchain>=0.3,<1.0
langchain-openai>=0.2,<1.0
langchain-chroma>=0.2,<1.0
langgraph>=0.2,<1.0
chromadb>=0.5,<1.0
tiktoken>=0.7,<1.0
